requirements.txt
discord.py==2.1.0
fastapi==0.88.0
python-dotenv==0.21.0
